pyproject.toml
[project]
name = "facemoment"
version = "0.1.0"
description = "Face moment detection and highlight clip extraction for 981park"
requires-python = ">=3.10"
dependencies = [
    "visualbase",
    "visualpath",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
]

[project.optional-dependencies]
# Development
dev = [
    "pytest>=7.0.0",
]

# IPC for distributed workers
zmq = [
    "pyzmq>=25.0.0",
]

# CLI/GUI visualization
cli = [
    "pyqt6>=6.10.2",
    "supervision>=0.19.0",
]

# === Worker-specific dependencies (isolated venvs) ===

# Face detection + expression analysis
face = [
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.16.0",
    "hsemotion-onnx>=0.3",
    "ml-dtypes>=0.5.0",
]

# Face with PyFeat (slower but more accurate expression)
face-full = [
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.16.0",
    "py-feat>=0.6.0",
    "pandas>=2.0.0,<3",
    "scipy>=1.10.0",
    "ml-dtypes>=0.5.0",
]

# Pose estimation
pose = [
    "ultralytics>=8.0.0",
]

# Hand gesture detection
gesture = [
    "mediapipe>=0.10.32",
]

# === Convenience bundles ===

# All ML backends (for development/testing - may have conflicts!)
all = [
    "facemoment[face,pose,gesture,cli]",
]

[project.scripts]
facemoment = "facemoment.cli:main"

# Register extractors and fusions as plugins for visualpath
[project.entry-points."visualpath.extractors"]
face = "facemoment.moment_detector.extractors.face:FaceExtractor"
pose = "facemoment.moment_detector.extractors.pose:PoseExtractor"
gesture = "facemoment.moment_detector.extractors.gesture:GestureExtractor"
quality = "facemoment.moment_detector.extractors.quality:QualityExtractor"
dummy = "facemoment.moment_detector.extractors.dummy:DummyExtractor"

[project.entry-points."visualpath.fusions"]
highlight = "facemoment.moment_detector.fusion.highlight:HighlightFusion"
dummy = "facemoment.moment_detector.fusion.dummy:DummyFusion"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/facemoment"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.uv.sources]
visualbase = { path = "../visualbase", editable = true }
visualpath = { path = "../visualpath", editable = true }

# NOTE: constraint/override dependencies removed!
# With VenvWorker, each extractor runs in an isolated venv,
# so dependency conflicts (opencv-python vs headless, onnxruntime vs gpu)
# are no longer an issue. See visualpath.process.VenvWorker.
